Update dependencies and enhance Android/iOS configurations
This commit updates various package versions in pubspec.lock, including downgrading the _fe_analyzer_shared and analyzer packages for compatibility. The Android build configuration is updated to use compileSdk version 36 and targetSdk version 36, while the Kotlin plugin version is upgraded to 2.1.0. Additionally, the iOS Podfile.lock is updated to reflect a new Flutter checksum, and the MinimumOSVersion in AppFrameworkInfo.plist is changed to 13.0 for better compatibility. Theme files are also modified to use BottomAppBarThemeData for improved theming consistency.