Update app configuration and assets for "Hopa" branding
This commit updates the app's package name and identifiers to "com.hopa.user" across multiple files, including AndroidManifest.xml and build.gradle. The smooth_page_indicator dependency is upgraded to version 1.2.1, and various onboarding SVG assets are removed in favor of new image assets. Additionally, the welcome description in the language file is revised for clarity. These changes aim to enhance the overall branding and user experience of the application.