Refactor navigation transitions and streamline restaurant routing
Removed the CustomPageTransition class to simplify navigation handling and replaced it with a default cupertino transition for improved performance. Updated the RestaurantsCardWidget to utilize a new method for opening restaurant screens with a 3D transition effect. Additionally, modified the RouteHelper to include a helper method for 3D transitions, enhancing the overall navigation experience. These changes aim to improve maintainability and user experience across the application.